// netrika/src/common/interfaces/NotCustomBlocksEnum.ts
export enum NotCustomBlocksEnum {
  MedicalCareCase = 1, // Случаи медицинского обслуживания
  MedicalOrganization = 2, // Медицинские организации
  AnamnesisOfLife = 3, // Анамнез жизни
  ObjectiveState = 4, // Объективное состояние
  ReferencePlan = 5, // План ведения пациента
  MedDocs = 6, //Медицинские документы
  AnamnesisOfDisease = 7, // Анамнез заболевания
  Diagnose = 8, // Диагноз
  Referrals = 9, // Направления
  UserBlock = 10, // Пользовательский блок
  InstrumentalAnalysis = 11, // Инструментальные исследования
  Aggregates = 12, // Расчётные показатели пациента
  Immunization = 13, // Иммунизация
  Documents = 14, // Документы
  PreferentialMedicationSupply = 17, // Льготное лекарственное обеспечение
  Dispensary = 18, // Диспансеризация
  Recipes = 19, // рецепты и льготы
  VaccinationCalendar = 20, // Календарь прививок
  TmReports = 21, //Телемедицинские консультации
  DynamicsLaboratoryResearch = 22, // Динамика лабораторных исследований
  MedicalServices = 23, // Медицинские услуги
}